pyproject.toml
[build-system]
requires = ["setuptools>=80.10.1"]
build-backend = "setuptools.build_meta"

[project]
name = "switchcraft"
version = "2026.1.4.dev0+d98c720"
description = "A universal silent switch finder and installer analyzer."
readme = "README.md"
authors = [{ name = "FaserF" }]
license = "MIT"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["installer", "silent-install", "automation", "winget", "msi", "exe"]
dependencies = [
    "pefile",
    "olefile",
    "requests",
    "click",
    "rich",
    "py7zr",
    "PyYAML",
    "keyring",
    "openai",
    "google-generativeai",
    "defusedxml",
    "PyJWT",
    "winotify",
    "anyio>=4.12.1"
]
requires-python = "==3.14"

[project.optional-dependencies]
gui = [
    "customtkinter",
    "tkinterdnd2",
    "pillow",
    "plyer"
]
modern = ["flet==0.80.2", "flet-desktop==0.80.2", "flet-charts==0.80.2"]
test = [
    "pytest==9.0.2",
    "flet==0.80.2",
    "flet-charts==0.80.2"
]
build = [
    "pyinstaller==6.18.0"
]

[project.scripts]
switchcraft = "switchcraft.main:main"

[project.urls]
Homepage = "https://github.com/FaserF/SwitchCraft"

[tool.setuptools.package-data]
switchcraft = ["assets/**/*", "assets/lang/*.json"]
switchcraft_winget = ["utils/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--ignore=test_out.txt --ignore=test_output.txt --ignore=test_addons_result.txt"

# Flet Web Publishing Configuration
[tool.flet.app]
path = "."
module = "main.py"
